Introspection: Move get_ports command from XDC plugin
